make test import generated lib
