--- test/main.cpp
@@ -1,13 +1,0 @@
-#include "DecisionTree.hpp"
-
-int main() {
-  Dataset d;
-  d.train.filename = "../test/data/san_fran_vs_nyc.csv";
-  d.test.filename = "../test/data/san_fran_vs_nyc_test.csv";
-
-  DecisionTree dt(d);
-  dt.print();
-  dt.test();
-  dt.generateGraph("../../graph.dot");
-  return 0;
-}
