+++ test/random_tester.cpp
@@ -1,0 +1,13 @@
+#include <DecisionTree/DecisionTree.hpp>
+
+int main() {
+  Dataset d;
+  d.train.filename = "data/san_fran_vs_nyc.csv";
+  d.test.filename = "data/san_fran_vs_nyc_test.csv";
+
+  DecisionTree dt(d);
+  dt.print();
+  dt.test();
+  dt.generateGraph("../../graph.dot");
+  return 0;
+}
